+++ examples/next-streaming-react-query/src/app/PrefetchBoundary/_components/ReactClientComponent.tsx
@@ -1,9 +1,0 @@
-'use client'
-
-import { useSuspenseQuery } from '@tanstack/react-query'
-import { query } from '~/query'
-
-export const ReactClientComponent = () => {
-  const { data } = useSuspenseQuery(query.text(1000))
-  return <div>{data}</div>
-}