+++ packages/react-query-4/src/PrefetchInfiniteQuery.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { type FetchInfiniteQueryOptions, type QueryKey } from '@tanstack/react-query'
 import { usePrefetchInfiniteQuery } from './usePrefetchInfiniteQuery'