+++ packages/react-query-5/src/index.ts
@@ -1,4 +1,5 @@
 export { infiniteQueryOptions } from './infiniteQueryOptions'
+export { QueriesHydration } from './QueriesHydration'
 export { queryOptions } from './queryOptions'
 export { usePrefetchInfiniteQuery } from './usePrefetchInfiniteQuery'
 export { usePrefetchQuery } from './usePrefetchQuery'