+++ packages/react-query-5/src/usePrefetchInfiniteQuery.ts
@@ -1,3 +1,5 @@
+'use client'
+
 import { usePrefetchInfiniteQuery as original_usePrefetchInfiniteQuery } from '@tanstack/react-query'
 
 /**