+++ packages/react-query/tsdown.config.ts
@@ -1,4 +1,10 @@
 import { options, scriptOptions } from '@suspensive/tsdown'
 import { defineConfig } from 'tsdown'
 
-export default defineConfig([options, scriptOptions])
+export default defineConfig([
+  {
+    ...options,
+    banner: { js: undefined },
+  },
+  scriptOptions,
+])